Final version "APS98-A0_01"

+++ computer_assignments/APS98-A0/APS98-A0_01.cpp
@@ -30,7 +30,7 @@ int main()
 
 string ibm_minus_one(string str)
 {
-  for (int i = 0; i < str.size(); i++)
+  for (int i = 0; i < (int)str.size(); i++)
   {
     str[i] != 'Z' ? str[i] = str[i] + 1 : str[i] = 'A';
   }
